+++ src/axom/slic/interface/slic_macros.hpp
@@ -299,7 +299,7 @@
 ///@}
 
 // Use complete debug macros when not on device
-#if defined(AXOM_DEBUG) && !defined(AXOM_DEVICE_CODE)
+#if (defined(AXOM_DEBUG) || defined(AXOM_ENABLE_SLIC_DEBUG_MACROS)) && !defined(AXOM_DEVICE_CODE)
 
   //-----------------------------------------------------------------------------
   /// @{
@@ -324,7 +324,9 @@
  * \param [in] EXP user-supplied boolean expression.
  *
  * \warning This macro calls processAbort() iff EXP is false.
- * \note This macro is only active when AXOM_DEBUG is defined.
+ * \note This macro is active when the AXOM_DEBUG_DEFINE CMake setting enables
+ *       AXOM_DEBUG (by default in Debug and RelWithDebInfo builds), or when
+ *       AXOM_ENABLE_SLIC_DEBUG_MACROS=ON.
  *
  * Usage:
  * \code
@@ -342,7 +344,9 @@
  * \param [in] msg user-supplied message
  *
  * \warning This macro calls processAbort() iff EXP is false.
- * \note This macro is only active when AXOM_DEBUG is defined.
+ * \note This macro is active when the AXOM_DEBUG_DEFINE CMake setting enables
+ *       AXOM_DEBUG (by default in Debug and RelWithDebInfo builds), or when
+ *       AXOM_ENABLE_SLIC_DEBUG_MACROS=ON.
  * \see SLIC_ASSERT( EXP )
  *
  * Usage:
@@ -399,7 +403,9 @@
  *  application is not aborted.
  *
  * \param [in] EXP user-supplied boolean expression.
- * \note This macro is only active when AXOM_DEBUG is defined.
+ * \note This macro is active when the AXOM_DEBUG_DEFINE CMake setting enables
+ *       AXOM_DEBUG (by default in Debug and RelWithDebInfo builds), or when
+ *       AXOM_ENABLE_SLIC_DEBUG_MACROS=ON.
  *
  * Usage:
  * \code
@@ -416,7 +422,9 @@
  * \param [in] EXP user-supplied boolean expression.
  * \param [in] msg user-supplied message
  *
- * \note This macro is only active when AXOM_DEBUG is defined.
+ * \note This macro is active when the AXOM_DEBUG_DEFINE CMake setting enables
+ *       AXOM_DEBUG (by default in Debug and RelWithDebInfo builds), or when
+ *       AXOM_ENABLE_SLIC_DEBUG_MACROS=ON.
  * \see SLIC_DEBUG( EXP )
  *
  * Usage:
@@ -456,7 +464,7 @@
 // Use assert when on device (note that messages are omitted).
 // Device HIP assert() tested with rocm@6.1.2
 // (ROCm support for device assert() begins with version 5.1.0).
-#elif defined(AXOM_DEBUG) && defined(AXOM_DEVICE_CODE)
+#elif (defined(AXOM_DEBUG) || defined(AXOM_ENABLE_SLIC_DEBUG_MACROS)) && defined(AXOM_DEVICE_CODE)
   #define SLIC_ASSERT(EXP) assert(EXP)
   #define SLIC_ASSERT_MSG(EXP, msg) assert(EXP)
   #define SLIC_CHECK(EXP) assert(EXP)
@@ -469,7 +477,7 @@
   #define SLIC_CHECK(ignore_EXP) ((void)0)
   #define SLIC_CHECK_MSG(ignore_EXP, ignore_msg) ((void)0)
 
-#endif /* END ifdef AXOM_DEBUG */
+#endif /* END if debug macros are enabled */
 
 /*!
  * \def SLIC_INFO( msg )
@@ -662,7 +670,7 @@
 #define SLIC_INFO_ROOT_IF_ONCE(EXP, msg) \
   SLIC_DETAIL_LOG_IF_ONCE(SLIC_INFO_IF, (EXP) && (axom::slic::isRoot()), msg)
 
-#ifdef AXOM_DEBUG
+#if defined(AXOM_DEBUG) || defined(AXOM_ENABLE_SLIC_DEBUG_MACROS)
 
   /*!
  * \def SLIC_DEBUG( msg )
@@ -670,7 +678,9 @@
  *
  * \param [in] msg user-supplied message
  *
- * \note The SLIC_Debug macro is active when AXOM_DEBUG is defined.
+ * \note This macro is active when the AXOM_DEBUG_DEFINE CMake setting enables
+ *       AXOM_DEBUG (by default in Debug and RelWithDebInfo builds), or when
+ *       AXOM_ENABLE_SLIC_DEBUG_MACROS=ON.
  *
  * Usage:
  * \code
@@ -686,7 +696,9 @@
  *
  * \param [in] msg user-supplied message
  *
- * \note The SLIC_DEBUG_ONCE macro is active when AXOM_DEBUG is defined.
+ * \note This macro is active when the AXOM_DEBUG_DEFINE CMake setting enables
+ *       AXOM_DEBUG (by default in Debug and RelWithDebInfo builds), or when
+ *       AXOM_ENABLE_SLIC_DEBUG_MACROS=ON.
  *
  * Usage:
  * \code
@@ -703,7 +715,9 @@
  * \param [in] EXP user-supplied boolean expression.
  * \param [in] msg user-supplied message.
  *
- * \note The SLIC_DEBUG_IF macro is active when AXOM_DEBUG is defined.
+ * \note This macro is active when the AXOM_DEBUG_DEFINE CMake setting enables
+ *       AXOM_DEBUG (by default in Debug and RelWithDebInfo builds), or when
+ *       AXOM_ENABLE_SLIC_DEBUG_MACROS=ON.
  *
  * Usage:
  * \code
@@ -729,7 +743,9 @@
  * \param [in] EXP user-supplied boolean expression.
  * \param [in] msg user-supplied message.
  *
- * \note The SLIC_DEBUG_IF_ONCE macro is active when AXOM_DEBUG is defined.
+ * \note This macro is active when the AXOM_DEBUG_DEFINE CMake setting enables
+ *       AXOM_DEBUG (by default in Debug and RelWithDebInfo builds), or when
+ *       AXOM_ENABLE_SLIC_DEBUG_MACROS=ON.
  *
  * Usage:
  * \code
@@ -745,7 +761,9 @@
  *
  * \param [in] msg user-supplied message.
  *
- * \note The SLIC_DEBUG_ROOT macro is active when AXOM_DEBUG is defined.
+ * \note This macro is active when the AXOM_DEBUG_DEFINE CMake setting enables
+ *       AXOM_DEBUG (by default in Debug and RelWithDebInfo builds), or when
+ *       AXOM_ENABLE_SLIC_DEBUG_MACROS=ON.
  *
  * Usage:
  * \code
@@ -761,7 +779,9 @@
  *
  * \param [in] msg user-supplied message.
  *
- * \note The SLIC_DEBUG_ROOT_ONCE macro is active when AXOM_DEBUG is defined.
+ * \note This macro is active when the AXOM_DEBUG_DEFINE CMake setting enables
+ *       AXOM_DEBUG (by default in Debug and RelWithDebInfo builds), or when
+ *       AXOM_ENABLE_SLIC_DEBUG_MACROS=ON.
  *
  * Usage:
  * \code
@@ -779,7 +799,9 @@
  * \param [in] EXP user-supplied boolean expression.
  * \param [in] msg user-supplied message.
  *
- * \note The SLIC_DEBUG_ROOT_IF macro is active when AXOM_DEBUG is defined.
+ * \note This macro is active when the AXOM_DEBUG_DEFINE CMake setting enables
+ *       AXOM_DEBUG (by default in Debug and RelWithDebInfo builds), or when
+ *       AXOM_ENABLE_SLIC_DEBUG_MACROS=ON.
  *
  * Usage:
  * \code
@@ -796,7 +818,9 @@
  * \param [in] EXP user-supplied boolean expression.
  * \param [in] msg user-supplied message.
  *
- * \note The SLIC_DEBUG_ROOT_IF_ONCE macro is active when AXOM_DEBUG is defined.
+ * \note This macro is active when the AXOM_DEBUG_DEFINE CMake setting enables
+ *       AXOM_DEBUG (by default in Debug and RelWithDebInfo builds), or when
+ *       AXOM_ENABLE_SLIC_DEBUG_MACROS=ON.
  *
  * Usage:
  * \code
@@ -815,7 +839,9 @@
  * \param [in] name The name of the container in the printed message.
  * \param [in] container The container (array, vector, view).
  *
- * \note The SLIC_DEBUG_PRINT_CONTAINER macro is active when AXOM_DEBUG is defined.
+ * \note This macro is active when the AXOM_DEBUG_DEFINE CMake setting enables
+ *       AXOM_DEBUG (by default in Debug and RelWithDebInfo builds), or when
+ *       AXOM_ENABLE_SLIC_DEBUG_MACROS=ON.
  *
  * Usage:
  * \code
@@ -840,7 +866,9 @@
  * \param [in] name The name of the container in the printed message.
  * \param [in] container The container (array, vector, view).
  *
- * \note The SLIC_DEBUG_PRINT_CONTAINER_ONCE macro is active when AXOM_DEBUG is defined.
+ * \note This macro is active when the AXOM_DEBUG_DEFINE CMake setting enables
+ *       AXOM_DEBUG (by default in Debug and RelWithDebInfo builds), or when
+ *       AXOM_ENABLE_SLIC_DEBUG_MACROS=ON.
  *
  * Usage:
  * \code